revise src dir
