+++ shardingsphere-control-panel/shardingsphere-cluster/shardingsphere-cluster-heartbeat/src/test/java/org/apache/shardingsphere/cluster/heartbeat/event/HeartBeatDetectNoticeEventTest.java
@@ -18,9 +18,6 @@
 package org.apache.shardingsphere.cluster.heartbeat.event;
 
 import org.junit.Test;
-import org.apache.shardingsphere.cluster.heartbeat.event.HeartBeatDetectNoticeEvent;
-import org.apache.shardingsphere.cluster.heartbeat.event.HeartBeatEventType;
-import org.apache.shardingsphere.cluster.heartbeat.event.HeartBeatDetectNoticeEvent;
 
 import static org.junit.Assert.assertTrue;
